requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-select2==8.3.0
requests==2.32.3